Версия 1.3.2 Доработан ГИС-модуль Реализация логики обработки ИО в очередях (таблица logistic) Добавлен тип данных "Дата-время (с миллисекундами)" и его обработка
git-svn-id: svn+ssh://svn.code.sf.net/p/dynamicdocs/code@839 05742d63-fc59-4ccd-b3f0-d7e2a2a8a88d

+++ ddocclient/kksfactory/KKSLoader.h
@@ -113,7 +113,7 @@ class _F_EXPORT KKSLoader
         
         KKSMap<qint64, KKSEIOData *> loadRecList (const KKSRubric * r) const;
         
-        KKSEIOData * loadEIOInfo (int idObject, int idRec) const;
+        KKSEIOData * loadEIOInfo (int idObject, qint64 idRec) const;
         
         //
         // просто данные в этих методах возвращаются в виде QList, а не QMap
